ai-orchestration/requirements.txt
# AI Orchestration Framework Dependencies
# All frameworks are MIT licensed - free for commercial use

# LangGraph - Master Orchestrator
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0

# CrewAI - Agent Teams
crewai>=0.80.0
crewai-tools>=0.14.0

# AutoGen - Conversational Control
autogen-agentchat>=0.4.0
autogen-ext[openai]>=0.4.0

# LlamaIndex - Knowledge & RAG
llama-index>=0.11.0
llama-index-llms-openai>=0.2.0
llama-index-embeddings-openai>=0.2.0
llama-index-vector-stores-supabase>=0.2.0

# Shared Dependencies
openai>=1.50.0
anthropic>=0.34.0
pydantic>=2.0.0
python-dotenv>=1.0.0
asyncio>=3.4.3
aiohttp>=3.9.0
httpx>=0.27.0

# Database & Storage
supabase>=2.0.0
psycopg2-binary>=2.9.0

# Utilities
tiktoken>=0.7.0
tenacity>=8.0.0
rich>=13.0.0
